requirements.txt
mcp>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
websockets>=15.0.0
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.16.0
